# swagger/docs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/booking/v1/service.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/booking/v1/bookings": {
      "post": {
        "operationId": "BookingService_CreateBooking",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateBookingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateBookingRequest"
            }
          }
        ],
        "tags": [
          "BookingService"
        ]
      }
    },
    "/api/booking/v1/bookings/{id}/cancel": {
      "post": {
        "operationId": "BookingService_CancelBooking",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CancelBookingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BookingServiceCancelBookingBody"
            }
          }
        ],
        "tags": [
          "BookingService"
        ]
      }
    },
    "/api/booking/v1/list-bookings": {
      "post": {
        "operationId": "BookingService_ListBookings",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListBookingsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListBookingsRequest"
            }
          }
        ],
        "tags": [
          "BookingService"
        ]
      }
    },
...
  },
  "definitions": {
    "BookingServiceCancelBookingBody": {
      "type": "object"
    },
...
    "bookingv1Page": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "string",
          "format": "uint64"
        },
        "offset": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Booking": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "workspace_id": {
          "type": "string",
          "format": "int64"
        },
        "user_id": {
          "type": "string",
          "format": "int64"
        },
        "start_at": {
          "type": "string",
          "format": "date-time"
        },
        "end_at": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "$ref": "#/definitions/v1BookingStatus"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1BookingStatus": {
      "type": "string",
      "enum": [
        "BOOKING_STATUS_UNSPECIFIED",
        "BOOKING_STATUS_ACTIVE",
        "BOOKING_STATUS_CANCELLED",
        "BOOKING_STATUS_COMPLETED"
      ],
      "default": "BOOKING_STATUS_UNSPECIFIED"
    },
    "v1CancelBookingResponse": {
      "type": "object"
    },
    "v1CreateBookingRequest": {
      "type": "object",
      "properties": {
        "workspace_id": {
          "type": "string",
          "format": "int64"
        },
        "user_id": {
          "type": "string",
          "format": "int64"
        },
        "start_at": {
          "type": "string",
          "format": "date-time"
        },
        "end_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1CreateBookingResponse": {
      "type": "object",
      "properties": {
        "booking_id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "v1ListBookingsRequest": {
      "type": "object",
      "properties": {
        "page": {
          "$ref": "#/definitions/bookingv1Page"
        },
        "filter": {
          "$ref": "#/definitions/v1ListBookingsRequestFilter"
        }
      }
    },
    "v1ListBookingsRequestFilter": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "format": "int64"
        },
        "workspace_id": {
          "type": "string",
          "format": "int64"
        },
        "statuses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1BookingStatus"
          }
        },
        "date_from": {
          "type": "string",
          "format": "date-time",
          "description": "date_from/date_to are truncated to a date on the server and matched\nagainst bookings whose [start_at, end_at) interval overlaps that\n[date_from, date_to] range of days."
        },
        "date_to": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1ListBookingsResponse": {
      "type": "object",
      "properties": {
        "bookings": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Booking"
          }
        },
        "total_count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
  }
}